+++ EF/Models/UserManagement/UmDivision.cs
@@ -15,7 +15,7 @@ public partial class UmDivision
 
     public bool IsActive { get; set; }
 
-    public long CreatedByUserId { get; set; }
+    public int CreatedByUserId { get; set; }
 
     public DateTime CreatedDate { get; set; }
